Allow null values in version- and changed_at-cols
This should make the generated objects less annoying to work with.

It allows the user more flexibility with setting not null for the
changed_at-column.

It also allows initializing version-columns with null. And also
allows saving objects without providing a version-attribute. I.e
the version attribute is only checked if the version value is
provided.